Add a machinery to workaround limitation with the module system:
Swift require modules, but currently, the SDK omit some useful headers from the declared modules, like asl, CommonCrypto, sqlite, and probably more.

To workaround that issue, we declare the missing modules in a specific module map and tell the swift compiler to lookup in the dir that contains that module map.

One problem with that approach is that a module map requires absolute paths, or paths relative to the map file. As we want to be able to point to different SDK, we have to create a map file per platform (actually, we use relative paths, and create a symlink to the corresponding SDK)